chore: move log models under buck2 types
