+++ api-model/src/lib.rs
@@ -1,4 +1,3 @@
 pub mod buck2;
 pub mod common;
 pub mod git;
-pub mod orion;
