--- api-model/src/orion/mod.rs
@@ -1,1 +1,0 @@
-pub mod log;
