+++ orion-server/src/log/log_service.rs
@@ -1,6 +1,6 @@
 use std::{path::Path, sync::Arc};
 
-use api_model::orion::log::LogEvent;
+use api_model::buck2::types::LogEvent;
 use futures::{Stream, StreamExt};
 use tokio_stream::wrappers::BroadcastStream;
