+++ orion-server/src/server.rs
@@ -44,16 +44,15 @@ use crate::{
     components(
         schemas(
             crate::scheduler::BuildRequest,
-            api_model::orion::log::LogSegment,
-            api_model::orion::log::LogLinesResponse,
-            api_model::orion::log::TargetLogLinesResponse,
-            api_model::orion::log::LogErrorResponse,
+            api_model::buck2::types::LogLinesResponse,
+            api_model::buck2::types::TargetLogLinesResponse,
+            api_model::buck2::types::LogErrorResponse,
             api::TaskStatusEnum,
             api::BuildDTO,
             api::TargetDTO,
-            api_model::orion::log::TargetLogQuery,
-            api_model::orion::log::LogReadMode,
-            api_model::orion::log::TaskHistoryQuery,
+            api_model::buck2::types::TargetLogQuery,
+            api_model::buck2::types::LogReadMode,
+            api_model::buck2::types::TaskHistoryQuery,
             api::TaskInfoDTO,
             api::OrionClientInfo,
             api::OrionClientStatus,
